defectsense-ai/defect-detection/requirements.txt
anomalib[cpu]==2.1.0
torch==2.10.0
torchvision==0.25.0
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
opencv-python==4.10.0.84
Pillow==11.0.0
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
alembic==1.14.0
pytest==8.3.3
httpx==0.27.2
scikit-learn==1.5.2
numpy==2.1.3
pyyaml==6.0.2
matplotlib==3.9.2
seaborn==0.13.2
tqdm==4.66.5
requests==2.32.3
